:sparkles: Define parameters as members of a special Enum class
The special class Param is a mix-in of the str type and the Enum class.
This mix-in allows the usage of the members of a subclass of Param
as common strings, useful as dictionary keys. In addition, the enum
members have a slot attribute that stores the order in which they were
defined.