🔨 🎨 ✅  Converting some attributes to properties in attrs classes
We want to take advantage of the attrs functions asdict and astuple to
convert a QMC model instance, as well as Sampling instances, to
dictionary objects, and then export the corresponding representation.
We have converted some non-initialization attributes to properties so
they are not included in the dictionary. Of course, we could use filters
in asdict (or in astuple), but we prefer not to have to.