🎨 🔨 ✅ Remove parameters num_batches and num_time_steps_batch from DMC Sampling constructor
The iterable interface of the DMC sampling has been changed so that it
yields states instead of batches, in principle, indefinitely. An
iterable object that yields batches of states configurations and
properties can be created through the ``batches`` method of a DMC
sampling instance.

In addition, the integer to seed the numba RNG is passed directly to the
states_generator.