✨ 🎨  Enhance the evaluation of estimators, as well as the calculation of their statistical errors
Several classes were introduced to handle the data and the calculation
of their statistical errors after a DMC sampling. It was added the option
to keep the data after sampling on memory. Also, it was added the
option to evaluate the mixed estimator of the structure factor instead
of the pure estimator.